backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-jose[cryptography]
passlib[bcrypt]
google-auth
google-auth-oauthlib
python-dotenv
pydantic
email-validator
requests
